update e2e tests and fixtures for JSON schema conversion pages
