+++ website/e2e/fixtures/pages/index.ts
@@ -6,4 +6,5 @@ export { ParsingPage as parsingPage } from "./parsing";
 export { ValidationPage as validationPage } from "./validation";
 export { StandardSchemaPage as standardSchemaPage } from "./standard";
 export { StringPage as stringPage } from "./string";
-export { JsonSchemaConversionPage as jsonSchemaConversionPage } from "./json-schema/conversion";
+export { ToJsonPage as toJsonPage } from "./json-schema/to-json";
+export { FromJsonPage as fromJsonPage } from "./json-schema/from-json";
