+++ website/e2e/fixtures/pages/json-schema/_conversion.ts
@@ -1,0 +1,35 @@
+import type { TableResult } from "@rickcedwhat/playwright-smart-table";
+import { useTable } from "@rickcedwhat/playwright-smart-table";
+
+import { trimSortLabels } from "#e2e/utils/index.ts";
+
+import { BasePage } from "../../base";
+
+export abstract class BaseConversionPage extends BasePage {
+  breakpoints = BasePage.defineBreakpoints({
+    desktop: ["laptop", "desktop"],
+  });
+
+  #tableHandle: TableResult<unknown> | undefined;
+  get desktop() {
+    const table = this.page.getByRole("table", { name: "Results" });
+    this.#tableHandle ??= useTable(table, {
+      headerTransformer: ({ text }) => trimSortLabels(text),
+    });
+    return {
+      table,
+      tableHandle: this.#tableHandle,
+    };
+  }
+
+  get mobile() {
+    const cardsList = this.page.getByRole("list", { name: "Results" });
+    const cards = cardsList.getByTestId("bench-card");
+    return {
+      cardsList,
+      cards,
+      getCardByLibraryName: (libraryName: string | RegExp) =>
+        cards.filter({ hasText: libraryName }),
+    };
+  }
+}
