+++ website/e2e/fixtures/pages/json-schema/from-json.ts
@@ -1,0 +1,5 @@
+import { BaseConversionPage } from "./_conversion";
+
+export class FromJsonPage extends BaseConversionPage {
+  url = "/json-schema/from-json";
+}
